Move custom derializer test classes to src/test
Relocated TargetClass, TargetClassDerializer, ClassUsingTarget, and
CustomDerializerTest to src/test/java/engine/serialization/custom/ to
keep src/main clean of test code. Updated package declarations and
imports accordingly.
