+++ nomad-realms/src/test/java/engine/serialization/custom/CustomDerializerTest.java
@@ -1,4 +1,4 @@
-package engine.serialization;
+package engine.serialization.custom;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
